improved properties handling and UX for assigning
